git-svn-id: svn+ssh://hgwdev.cse.ucsc.edu/projects/compbio/svnroot/hausslerlab/recon/branches/2009-09-16/src/sonLib@5530 a8237c23-d22b-0410-ac1a-db40f18c2caa

+++ impl/commonC.c
@@ -537,7 +537,7 @@ int hashtable_longEqualKey( const void *key1, const void *key2 ) {
 }
 
 uint32_t hashtable_key( const void *k ) {
-	return (uint32_t)k;
+	return (uint32_t)(size_t)k;
 }
 
 int hashtable_equalKey( const void *key1, const void *key2 ) {
